Create Notice
Hi everyone, this is killjoy37. If you are reading this and are at least a little confused about GitHub, believe me, you are not alone. Despite what I had originally thought, GitHub has nothing in place that would allow us to chat in real time. So, to communicate we must send smoke sig-- Oh, wait, there is a thing called Wiki where we can post pages of things. Great! Also, I presume that creating an Issue is much like creating a thread in Roblox's forums. Keep on the lookout for new issues and wiki's because I'll figure out how to post one soon. Thanks!